Finished the Tournament API
made the verbose option use "pretty print"

+++ src/main/java/no/stelar7/api/l4j8/pojo/tournament/TournamentCode.java
@@ -5,7 +5,7 @@ import java.util.Set;
 public class TournamentCode
 {
     String    code;
-    Integer   id;
+    Long      id;
     String    lobbyName;
     String    map;
     String    metaData;
@@ -29,7 +29,7 @@ public class TournamentCode
         return code;
     }
 
-    public Integer getId()
+    public Long getId()
     {
         return id;
     }
